Move server code to server folder
