+++ server/utils/parseDocumentIds.ts
@@ -1,5 +1,5 @@
 import { Node } from "prosemirror-model";
-import { parser } from "@shared/editor";
+import { parser } from "@server/editor";
 
 export default function parseDocumentIds(text: string): string[] {
   const value = parser.parse(text);
